+++ test/common.go
@@ -103,7 +103,7 @@ func CreateTestInput(t *testing.T, isCheckTx bool) (sdk.Context, staking.Keeper,
 func GenRandCheckpointHeader(start int, headerSize int) (headerBlock types.CheckpointBlockHeader, err error) {
 	start = start
 	end := start + headerSize
-	maxCheckpointLenght := uint(1024)
+	maxCheckpointLenght := uint64(1024)
 	roothash, err := checkpointTypes.GetHeaders(uint64(start), uint64(end), maxCheckpointLenght)
 	if err != nil {
 		return headerBlock, err
